wyq_整合

+++ src/main/java/com/capgemini/cn/demo/userSystem/entity/BranchInfo.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Data
 public class BranchInfo {
 	
-	private int branchId;
+	private Long branchId;
 	
 	private String branchName;
